fix forbidden

+++ app/Http/Controllers/CheckoutController.php
@@ -121,7 +121,7 @@ class CheckoutController extends Controller
 
     public function success(Order $order)
     {
-        if ($order->user_id !== Auth::id()) {
+        if ((int) $order->user_id !== (int) Auth::id()) {
             abort(403);
         }
